+++ ProjectPostmail/Receiver.cs
@@ -8,11 +8,11 @@ namespace ProjectPostmail
 {
     class Receiver : Person
     {
-        private readonly float _sender_id;
+        private readonly double _sender_id;
         private readonly int _postoffice_number;
         private readonly double _price;
 
-        public Receiver(int age, string name, string surname, float id, float sender_id, int postoffice_number, double price) : base(age, name, surname, id)
+        public Receiver(int age, string name, string surname, double id, double sender_id, int postoffice_number, double price) : base(age, name, surname, id)
         {
             _sender_id = sender_id;
             _postoffice_number = postoffice_number;
@@ -24,7 +24,7 @@ namespace ProjectPostmail
             return base.GetInfo() + $"Sender id: {_sender_id}, Postoffice number: {_postoffice_number}";
         }
 
-        public float GetSenderID()
+        public double GetSenderID()
         {
             return _sender_id;
         }
